ai/requirements.txt
absl-py==2.2.2
antlr4-python3-runtime==4.9.3
anytree==2.13.0
asttokens==3.0.0
babel==2.17.0
backrefs==5.8
build==1.2.2.post1
certifi==2025.1.31
charset-normalizer==3.4.1
check-manifest==0.50
click==8.2.0
colorama==0.4.6
contourpy==1.3.1
coverage==7.8.0
cycler==0.12.1
decorator==5.2.1
easydict==1.13
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
ghp-import==2.1.0
gitdb==4.0.12
GitPython==3.1.44
griffe==1.7.3
grpcio==1.71.0
hydra-core==1.3.2
idna==3.10
iniconfig==2.1.0
ipython==8.36.0
jedi==0.19.2
Jinja2==3.1.6
kafka-python==2.1.5
kiwisolver==1.4.8
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-autorefs==1.4.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.14
mkdocs-material-extensions==1.3.1
mkdocstrings==0.29.1
mkdocstrings-python==1.16.10
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.1
omegaconf==2.3.0
opencv-python==4.11.0.86
packaging==24.2
paginate==0.5.7
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.5.0
prompt_toolkit==3.0.51
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
Pygments==2.19.1
pymdown-extensions==10.15
pyparsing==3.2.3
pyproject_hooks==1.2.0
pytest==8.3.5
pytest-cov==6.1.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyyaml_env_tag==1.1
requests==2.32.3
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
thop==0.1.1.post2209072238
tomli==2.2.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
-e git+https://github.com/kookmin-sw/capstone-2025-10.git@1ce70fdf3baa4652d73af3c5acf18e819fb0b0f7#egg=ultralytics&subdirectory=ai/src/YOLOv8-DeepSORT-Object-Tracking
ultralytics-thop==2.0.14
urllib3==2.4.0
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
